changed the roles that can use delete

+++ commands/mdelete.js
@@ -9,11 +9,16 @@ module.exports = {
                 .setDescription('Set the amount of messages you want to delete')
                 .setMinValue(1)
                 .setRequired(true)
+        )
+        .addUserOption(option =>
+            option.setName('user')
+                .setDescription('Specify a user to delete messages from')
         ),
 
     async execute(interaction) {
         const number = interaction.options.getInteger('number');
-        const user = interaction.options.getString('user');
+        const user = interaction.options.getUser('user');
+        const userMentioned = interaction.member;
 
         const row = new ActionRowBuilder()
             .addComponents(
@@ -27,36 +32,45 @@ module.exports = {
                     .setStyle(ButtonStyle.Secondary)
             );
 
-        await interaction.reply({ content: `Are you sure you want to delete ${number} messages${user ? ` from ${user}` : ''}?`, components: [row], ephemeral: true });
+        const goodRoles = ['632250692509237268', '632244499292225583', '632244879216345138'];
+        const hasGoodRole = goodRoles.some(role => userMentioned.roles.cache.has(role));
+        if (!hasGoodRole) {
+            return interaction.reply({
+                content: `You can't delete messages because you don't have a ${goodRoles.map(role => `<@&${role}>`).join(' or ')}`,
+                ephemeral: true,
+            });
+        } else {
+            await interaction.reply({ content: `Are you sure you want to delete ${number} messages${user ? ` from ${user.username}` : ''}?`, components: [row], ephemeral: true });
 
-        const filter = i => i.customId === 'confirm' || i.customId === 'cancel';
-        const collector = interaction.channel.createMessageComponentCollector({ filter, time: 15000 });
+            const filter = i => i.customId === 'confirm' || i.customId === 'cancel';
+            const collector = interaction.channel.createMessageComponentCollector({ filter, time: 15000 });
 
-        collector.on('collect', async i => {
-            if (i.customId === 'confirm') {
-                let messages = await interaction.channel.messages.fetch({ limit: number });
+            collector.on('collect', async i => {
+                if (i.customId === 'confirm') {
+                    let messages = await interaction.channel.messages.fetch({ limit: number });
 
-                if (user) {
-                    messages = messages.filter(msg => msg.author.username === user);
+                    if (user) {
+                        messages = messages.filter(msg => msg.author.id === user.id);
+                    }
+
+                    await interaction.channel.bulkDelete(messages, true)
+                        .then(deletedMessages => {
+                            i.update({ content: `Successfully deleted ${deletedMessages.size} messages.`, components: [] });
+                        })
+                        .catch(error => {
+                            console.error('Error deleting messages:', error);
+                            i.update({ content: 'There was an error trying to delete messages in this channel.', components: [] });
+                        });
+                } else if (i.customId === 'cancel') {
+                    i.update({ content: 'Message deletion canceled.', components: [] });
                 }
+            });
 
-                await interaction.channel.bulkDelete(messages, true)
-                    .then(deletedMessages => {
-                        i.update({ content: `Successfully deleted ${deletedMessages.size} messages.`, components: [] });
-                    })
-                    .catch(error => {
-                        console.error('Error deleting messages:', error);
-                        i.update({ content: 'There was an error trying to delete messages in this channel.', components: [] });
-                    });
-            } else if (i.customId === 'cancel') {
-                i.update({ content: 'Message deletion canceled.', components: [] });
-            }
-        });
-
-        collector.on('end', collected => {
-            if (collected.size === 0) {
-                interaction.editReply({ content: 'Message deletion timed out (were you sleeping or what?).', components: [] });
-            }
-        });
+            collector.on('end', collected => {
+                if (collected.size === 0) {
+                    interaction.editReply({ content: 'Message deletion timed out (were you sleeping or what?).', components: [] });
+                }
+            });
+        }
     },
 };
